Env file created
