+++ lib/env.dart
@@ -1,0 +1,11 @@
+class Env {
+  static final String appname = 'DocLense';
+
+  static final String wiredashID = 'doclense-sr0pzs4';
+
+  static final String wiredashSecret =
+      '0f4ajalausiiv0i7po2outfhtvmxy9a6dcd1o9rqk3a3ibcn';
+}
+
+
+// static final String a = '';
